requirements.txt
# Django и Django REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3

# CORS headers
django-cors-headers==4.3.0

# Swagger/OpenAPI документация
drf-yasg==1.21.7

# База данных
dj-database-url==2.1.0

# Сервер для продакшена
gunicorn==21.2.0

# Тестирование
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Утилиты
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0

setuptools==80.9.0